added comments and doc strings to the auth.py file and init file in views. removed duplicate imports and moved user and post serializers to their respective momdules
