+++ rareapi/views/__init__.py
@@ -1,3 +1,4 @@
+"""Rare package"""
 from .auth import login_user
 from .auth import register_user
 from .tag import Tags
@@ -6,4 +7,3 @@ from .post import Posts
 from .comment import Comments
 from .category import Categories
 from .user import Users
-# from .rareuser import RareUsers
