+++ rareapi/views/auth.py
@@ -1,3 +1,4 @@
+"""View module for handling authentication and new user registration"""
 import json
 import datetime
 from django.http import HttpResponse
@@ -8,9 +9,7 @@ from rest_framework import status
 from rest_framework.authtoken.models import Token
 from rareapi.models import RareUser
 
-
 @csrf_exempt
-
 def login_user(request):
     '''Handles the authentication of a user
     Method arguments:
@@ -22,25 +21,23 @@ def login_user(request):
     # If the request is a HTTP POST, try to pull out the relevant information.
     if request.method == 'POST':
 
-        # Use the built-in authenticate method to verify
         username = req_body['username']
         password = req_body['password']
         authenticated_user = authenticate(username=username, password=password)
 
-        # If authentication was successful, respond with their token
+        # If authentication was successful, respond with true and the users token
         if authenticated_user is not None:
             token = Token.objects.get(user=authenticated_user)
             rare_user = RareUser.objects.get(user=authenticated_user)
-            data = json.dumps({"valid": True, "token": token.key, "is_staff": rare_user.is_staff})
+            data = json.dumps({"valid": True, "token": token.key})
             return HttpResponse(data, content_type='application/json')
 
         else:
-            # Bad login details were provided. So we can't log the user in.
+            # if authentication fails, return false with no token
             data = json.dumps({"valid": False})
             return HttpResponse(data, content_type='application/json')
 
 @csrf_exempt
-
 def register_user(request):
     '''Handles the creation of a new user for authentication
     Method arguments:
@@ -50,8 +47,7 @@ def register_user(request):
     # Load the JSON string of the request body into a dict
     req_body = json.loads(request.body.decode())
 
-    # Create a new user by invoking the `create_user` helper method
-    # on Django's built-in User model
+    # Create a new user by invoking the `create_user` helper method on Django's built-in User model
     new_user = User.objects.create_user(
         username=req_body['username'],
         email=req_body['email'],
@@ -61,14 +57,14 @@ def register_user(request):
         is_active=True,
         is_staff=False
     )
-
+    #rare_user has a property `user` which makes all `user` properties accessible through the rare_user
     rare_user = RareUser.objects.create(
         bio=req_body['bio'],
         profile_image_url=req_body['profile_image_url'],
         user=new_user
     )
 
-    # Commit the user to the database by saving it
+    # save it all to the db
     rare_user.save()
 
     # Use the REST Framework's token generator on the new user account
